Split out TypeList into its own header and moved pragma macros into Platform.h
